+++ dashboard-app/src/components/StackedBarChart/Chart.tsx
@@ -10,22 +10,23 @@ import { getStackedGraphOptions, totalizer } from './utils/chartjsUtils';
 import _Card from '../Card';
 import { H3 } from '../Headings';
 import { ColoursEnum } from '../../styles/design_system';
+import { Paragraph } from '../Typography';
 
 
 /*
  * Types
  */
-
 interface Props {
   data: ChartData<chartjs.ChartData>;
   xAxisTitle: string;
   yAxisTitle: string;
   title: string;
 }
+
+
 /*
  * Styles
  */
-
 const MainCol = styled(Col)`
   padding: 0 !important;
 `;
@@ -40,15 +41,46 @@ const Title = styled(H3)`
   text-align: left;
 `;
 
-/*
+const NoDataContainer = styled.div`
+  height: 10rem;
+  width: 100%;
+  color: ${ColoursEnum.white};
+  display: flex;
+  align-items: center;
+  justify-content: center;
+`;
+
+
+/**
+ * Helpers
+ */
+const datasetKeyProvider = (d: { id: number }) => d.id;
+const checkIsDataEmpty = (cd: any) => (!cd || cd.datasets.length === 0);
+
+
+/**
  * Components
  */
+const noData = (
+  <Col xs={12}>
+    <NoDataContainer>
+      <Paragraph>NO DATA AVAILABLE</Paragraph>
+    </NoDataContainer>
+  </Col>
+);
 
 const Chart: FunctionComponent<Props> = (props) => {
   const { data, xAxisTitle, yAxisTitle, title } = props;
-  const datasetKeyProvider = (d: any) => d.id;
+  const graph = (
+    <RoundedBar
+      datasetKeyProvider={datasetKeyProvider}
+      plugins={[totalizer, ChartDataLabels]}
+      data={data}
+      options={getStackedGraphOptions(xAxisTitle, yAxisTitle)}
+    />
+  );
+
   return (
-  <MainCol xs={9}>
     <Card>
       <Row center="xs">
         <Col xs={12}>
@@ -57,16 +89,15 @@ const Chart: FunctionComponent<Props> = (props) => {
       </Row>
       <Row center="xs" middle="xs">
         <Col xs={12} md={9}>
-        <RoundedBar
-          datasetKeyProvider={datasetKeyProvider}
-          plugins={[totalizer, ChartDataLabels]}
-          data={data}
-          options={getStackedGraphOptions(xAxisTitle, yAxisTitle)}
-        />
+          {
+            checkIsDataEmpty(data)
+              ? noData
+              : graph
+          }
         </Col>
       </Row>
     </Card>
-  </MainCol>);
+  );
 };
 
 export default Chart;
