+++ dashboard-app/src/dashboard/ByTime/TimeTabs.tsx
@@ -1,18 +1,17 @@
 import React, { FunctionComponent } from 'react';
-import styled from 'styled-components';
 import { Row, Col } from 'react-flexbox-grid';
 
-import _DataTable from '../../components/DataTable';
+import DataTable from '../../components/DataTable';
 import { TabGroup } from '../../components/Tabs';
 import { AggregatedData } from '../dataManipulation/logsToAggregatedData';
 import { DurationUnitEnum } from '../../types';
 import { TableData } from '../dataManipulation/aggregatedToTableData';
 import StackedBarChart from '../../components/StackedBarChart/index';
 
+
 /*
  * Types
  */
-
 interface Props {
   data?: AggregatedData;
   unit: DurationUnitEnum;
@@ -22,11 +21,6 @@ interface Props {
   title: string;
 }
 
-/*
- * Styles
- */
-const DataTable = styled(_DataTable)`
-`;
 
 /*
  * Component
@@ -34,7 +28,7 @@ const DataTable = styled(_DataTable)`
 const TimeTabs: FunctionComponent<Props> = (props) => {
   const { data, unit, tableData, sortBy, onChangeSortBy, title } = props;
 
-  return(
+  return (
     <Row center="xs">
       <Col xs={12}>
         <TabGroup titles={['Chart', 'Table']}>
